reworked cooldown and message parsing
Previously the bot did a cooldown on the messages it parses. Each command was defined at two places: The place handling the command and the place doing the cooldown.

With this change the bot does the cooldown on its own outgoing messages. Each command is processed no matter if the bot is in cooldown or not, just the output perhaps get supressed. Because spamming with many commands in a single message is not possible anymore it's not necessary anymore to limit the number of processed commands per message.